PORT-12830-cleanup
