+++ port/folder/model.go
@@ -5,8 +5,7 @@ import "github.com/hashicorp/terraform-plugin-framework/types"
 type FolderModel struct {
 	ID         types.String `tfsdk:"id"`
 	Identifier types.String `tfsdk:"identifier"`
-	// Sidebar    types.String `tfsdk:"sidebar"`
-	Title  types.String `tfsdk:"title"`
-	After  types.String `tfsdk:"after"`
-	Parent types.String `tfsdk:"parent"`
+	Title      types.String `tfsdk:"title"`
+	After      types.String `tfsdk:"after"`
+	Parent     types.String `tfsdk:"parent"`
 }
