+++ port/folder/refreshFolderToState.go
@@ -8,7 +8,7 @@ import (
 func refreshFolderToState(fm *FolderModel, f *cli.Folder) error {
 	fm.Identifier = types.StringValue(f.Identifier)
 	fm.ID = types.StringValue(f.Identifier)
-	// fm.Sidebar = types.StringValue(f.Sidebar)
+
 	if f.Title != "" {
 		fm.Title = types.StringValue(f.Title)
 	}
@@ -21,8 +21,5 @@ func refreshFolderToState(fm *FolderModel, f *cli.Folder) error {
 		fm.Parent = types.StringValue(f.Parent)
 	}
 
-	// fm.Title = types.StringValue(f.Title)
-	// fm.After = types.StringValue(f.After)
-	// fm.Parent = types.StringValue(f.Parent)
 	return nil
 }
