+++ port/folder/schema.go
@@ -1,6 +1,10 @@
 package folder
 
 import (
+	"context"
+	"os"
+
+	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
 )
 
@@ -29,27 +33,27 @@ func FolderSchema() map[string]schema.Attribute {
 	}
 }
 
-// func (r *FolderResource) Schema(ctx context.Context, req resource.SchemaRequest, resp *resource.SchemaResponse) {
-// 	resp.Schema = schema.Schema{
-// 		MarkdownDescription: FolderResourceMarkdownDescription,
-// 		Attributes:          FolderSchema(),
-// 	}
-// }
+func (r *FolderResource) Schema(ctx context.Context, req resource.SchemaRequest, resp *resource.SchemaResponse) {
+	resp.Schema = schema.Schema{
+		MarkdownDescription: FolderResourceMarkdownDescription,
+		Attributes:          FolderSchema(),
+	}
+}
 
-// func (r *FolderResource) ValidateConfig(ctx context.Context, req resource.ValidateConfigRequest, resp *resource.ValidateConfigResponse) {
-// 	var state FolderModel
-// 	resp.Diagnostics.Append(req.Config.Get(ctx, &state)...)
+func (r *FolderResource) ValidateConfig(ctx context.Context, req resource.ValidateConfigRequest, resp *resource.ValidateConfigResponse) {
+	var state FolderModel
+	resp.Diagnostics.Append(req.Config.Get(ctx, &state)...)
 
-// 	if resp.Diagnostics.HasError() {
-// 		return
-// 	}
+	if resp.Diagnostics.HasError() {
+		return
+	}
 
-// 	betaFeaturesEnabledEnv := os.Getenv("PORT_BETA_FEATURES_ENABLED")
-// 	if !(betaFeaturesEnabledEnv == "true") {
-// 		resp.Diagnostics.AddError("Beta features are not enabled", "Folder resource is currently in beta and is subject to change in future versions. Use it by setting the Environment Variable PORT_BETA_FEATURES_ENABLED=true.")
-// 		return
-// 	}
-// }
+	betaFeaturesEnabledEnv := os.Getenv("PORT_BETA_FEATURES_ENABLED")
+	if !(betaFeaturesEnabledEnv == "true") {
+		resp.Diagnostics.AddError("Beta features are not enabled", "Folder resource is currently in beta and is subject to change in future versions. Use it by setting the Environment Variable PORT_BETA_FEATURES_ENABLED=true.")
+		return
+	}
+}
 
 var FolderResourceMarkdownDescription = `
